Created HeaderForUserElement refactored existing code
