+++ TreshchovProj/src/main/java/org/pages/HomePage.java
@@ -1,38 +1,27 @@
 package org.pages;
 
-import org.apache.log4j.Logger;
-import org.junit.Assert;
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
+import org.pages.elements.HeaderForUserElement;
 
 public class HomePage extends ParentPage {
-    Logger logger = Logger.getLogger(getClass());
+  //  Logger logger = Logger.getLogger(getClass());
 
-    @FindBy(xpath = "//button[text()='Sign Out']")
-    private WebElement buttonSignOut;
 
-    @FindBy(xpath = "//a[@href='/create-post']")
-    private WebElement buttonCreatePost;
 
     public HomePage(WebDriver webDriver) {
         super(webDriver);
     }
 
-    public void checkIsButtonSignOutVisible() {
-        checkIsElementVisible(buttonSignOut);
+    public HeaderForUserElement getHeaderElement() {
+        return new HeaderForUserElement(webDriver);
     }
 
     public HomePage checkIsRedirectToHomePage() {
-        checkIsButtonSignOutVisible();
+        getHeaderElement().checkIsButtonSignOutVisible();
         //TODO check current URL
 
         return this;
     }
 
-    public CreateNewPostPage clickOnButtonCreatePost(){
-        clickOnElement(buttonCreatePost);
-        return new CreateNewPostPage(webDriver);
-    }
+
 }
