+++ TreshchovProj/src/main/java/org/pages/MyProfilePage.java
@@ -1,0 +1,9 @@
+package org.pages;
+
+import org.openqa.selenium.WebDriver;
+
+public class MyProfilePage extends ParentPage {
+    public MyProfilePage(WebDriver webDriver) {
+        super(webDriver);
+    }
+}
