+++ TreshchovProj/src/main/java/org/pages/PageProvider.java
@@ -17,4 +17,9 @@ public class PageProvider {
     public HomePage getHomePage() {
         return new HomePage(webDriver);
     }
+
+    public PostPage getPostPage() {
+        return new PostPage(webDriver);
+    }
+
 }
