+++ TreshchovProj/src/test/java/org/LoginTest/LoginTestWithPageObject.java
@@ -15,7 +15,7 @@ public class LoginTestWithPageObject extends BaseTest {
                 enterTextIntoInputPassword(VALID_PASSWORD).
                 clickOnButtonSignIn();
 
-        pageProvider.getHomePage().checkIsButtonSignOutVisible();
+        pageProvider.getHomePage().getHeaderElement().checkIsButtonSignOutVisible();
 
 
     }
